Update authentication templates
