+++ authentication-typescript/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## 1.0.0
+## v1.1.0
+
+### Changed
+-   Updated dependencies to use latest packages.
+-   Reorganized folder structure
+
+### Fixed
+-   Jest configuration allows default tests to pass.
+
+
+## v1.0.0
 
 ### Added
