+++ authentication-typescript/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@pxblue/react-native-template-authentication-typescript",
-    "version": "1.0.0",
+    "version": "1.1.0",
     "keywords": [
         "react native",
         "template",
