+++ authentication-typescript/template/src/constants/index.ts
