+++ authentication-typescript/template/src/constants/sampleEula.ts
