+++ authentication-typescript/template/src/navigation/DeepLinking.ts
