+++ authentication-typescript/template/src/navigation/index.tsx
@@ -1,27 +1,34 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import { View } from 'react-native';
-import { NavigationDrawer } from './navigation-drawer';
+import { NavDrawerProps, NavigationDrawer } from './navigation-drawer';
 import { createStackNavigator } from '@react-navigation/stack';
-import Home from '../pages/home';
-import PageOne from '../pages/pageOne';
-import PageTwo from '../pages/pageTwo';
+import Home from '../screens/home';
+import PageOne from '../screens/pageOne';
+import PageTwo from '../screens/pageTwo';
 
 const Drawer = createDrawerNavigator();
 
-const RootStack = createStackNavigator();
+export type RootStackParamList = {
+    Home: undefined;
+    PageOne: undefined;
+    PageTwo: undefined;
+    NavigationDrawer: undefined;
+};
 
-const CustomDrawerContent = (props) => (
+const RootStack = createStackNavigator<RootStackParamList>();
+
+const CustomDrawerContent = (props: any): any => (
     <View style={{ height: '100%' }}>
         <NavigationDrawer {...props} />
     </View>
 );
 
-export const MainRouter = () => (
+export const MainRouter = (): any => (
     <Drawer.Navigator
         initialRouteName="Home"
         drawerStyle={{ backgroundColor: 'transparent', width: '80%' }}
-        drawerContent={(props) => <CustomDrawerContent {...props} />}
+        drawerContent={(props: NavDrawerProps): ReactNode => <CustomDrawerContent {...props} />}
     >
         <RootStack.Screen name="Home" component={Home} />
         <RootStack.Screen name="PageOne" component={PageOne} />
