+++ authentication-typescript/template/src/screens/index.tsx
