+++ authentication-typescript/template/src/screens/pageOne.tsx
@@ -3,10 +3,10 @@ import { SafeAreaView, ScrollView, StyleSheet, ViewStyle } from 'react-native';
 import { EmptyState, Header, InfoListItemProps, UserMenu, wrapIcon } from '@pxblue/react-native-components';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../router';
-import { useSecurityActions } from '@pxblue/react-native-auth-workflow';
-import { LocalStorage } from '../store/local-storage';
+import { RootStackParamList } from '../navigation';
 import { Avatar } from 'react-native-paper';
+import { LocalStorage } from '../store/local-storage';
+import { useSecurityActions } from '@pxblue/react-native-auth-workflow';
 import * as Colors from '@pxblue/colors';
 
 const Event = wrapIcon({ IconClass: MatIcon, name: 'event', flip: false });
@@ -30,10 +30,10 @@ const styles = (): StyleSheet.NamedStyles<{
     });
 
 type AppProps = {
-    navigation: StackNavigationProp<RootStackParamList, 'PageTwo'>;
+    navigation: StackNavigationProp<RootStackParamList, 'PageOne'>;
 };
 
-const PageTwo: React.FC<AppProps> = ({ navigation }): JSX.Element => {
+const PageOne: React.FC<AppProps> = ({ navigation }): JSX.Element => {
     const defaultStyles = styles();
     const securityHelper = useSecurityActions();
 
@@ -54,7 +54,7 @@ const PageTwo: React.FC<AppProps> = ({ navigation }): JSX.Element => {
     return (
         <>
             <Header
-                title={'Page Two'}
+                title={'Page One'}
                 navigation={{
                     icon: MenuIcon,
                     onPress: (): void => {
@@ -92,4 +92,4 @@ const PageTwo: React.FC<AppProps> = ({ navigation }): JSX.Element => {
     );
 };
 
-export default PageTwo;
+export default PageOne;
