+++ authentication-typescript/template/src/store/local-storage.ts
