+++ authentication/template/src/actions/RegistrationUIActions.js
