+++ authentication/template/src/constants/index.js
