+++ authentication/template/src/constants/sampleEula.js
