+++ authentication/template/src/screens/index.js
