+++ authentication/template/src/store/local-storage.js
