+++ blank-typescript/dependencies.json
@@ -8,6 +8,7 @@
         "@pxblue/icons-svg@^1.2.0",
         "react-native-vector-icons@^8.0.0",
         "react-native-svg@^12.1.0",
+        "react-native-gesture-handler@^1.10.2",
         "react-native-svg-transformer@^0.14.3",
         "react-native-safe-area-context@^3.2.0"
     ],
