+++ routing-typescript/template/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider as ThemeProvider } from 'react-native-paper';
 import * as PXBThemes from '@pxblue/react-native-themes';
-import { MainRouter } from './navigation';
+import { MainRouter } from './src/navigation';
 
 export const App = (): JSX.Element => (
     <ThemeProvider theme={PXBThemes.blue}>
