+++ routing/template/src/screens/home.js
@@ -1,42 +1,13 @@
 import React, { useCallback } from 'react';
-import {
-    SafeAreaView,
-    ScrollView,
-    StyleSheet,
-    Linking,
-    TextStyle,
-    ViewStyle,
-    View,
-    Animated,
-    Easing,
-} from 'react-native';
-import { Avatar, Button, Divider, useTheme } from 'react-native-paper';
-import { Body1, H4, Header, InfoListItemProps, UserMenu, wrapIcon } from '@pxblue/react-native-components';
-import { Theme } from 'react-native-paper/lib/typescript/types';
-import Logo from '../assets/images/Logo.svg';
+import { SafeAreaView, ScrollView, StyleSheet, Linking, View, Animated, Easing } from 'react-native';
+import { Button, Divider, useTheme } from 'react-native-paper';
+import { Body1, H4, Header, wrapIcon } from '@pxblue/react-native-components';
+import Logo from '../../assets/images/Logo.svg';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
-import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../router';
-import { useSecurityActions } from '@pxblue/react-native-auth-workflow';
-import { LocalStorage } from '../store/local-storage';
-import * as Colors from '@pxblue/colors';
 
 const MenuIcon = wrapIcon({ IconClass: MatIcon, name: 'menu', flip: false });
-const LockIcon = wrapIcon({ IconClass: MatIcon, name: 'lock', flip: false });
-const ExitToAppIcon = wrapIcon({ IconClass: MatIcon, name: 'exit-to-app', flip: false });
 
-const styles = (
-    theme: Theme
-): StyleSheet.NamedStyles<{
-    content: ViewStyle;
-    pxbLogoWrapper: ViewStyle;
-    pxbLogo: ViewStyle;
-    title: TextStyle;
-    subtitle: TextStyle;
-    bold: TextStyle;
-    divider: ViewStyle;
-    openURLButtonText: TextStyle;
-}> =>
+const styles = (theme) =>
     StyleSheet.create({
         content: {
             flex: 1,
@@ -69,7 +40,7 @@ const styles = (
         },
     });
 
-const OpenURLButton = (props: any): JSX.Element => {
+const OpenURLButton = (props) => {
     const { url, title } = props;
     const theme = useTheme();
     const defaultStyles = styles(theme);
@@ -79,25 +50,16 @@ const OpenURLButton = (props: any): JSX.Element => {
     }, [url]);
 
     return (
-        <Button
-            onPress={(): Promise<void> => handlePress()}
-            labelStyle={defaultStyles.openURLButtonText}
-            uppercase={false}
-        >
+        <Button onPress={() => handlePress()} labelStyle={defaultStyles.openURLButtonText} uppercase={false}>
             {title}
         </Button>
     );
 };
 
-type AppProps = {
-    navigation: StackNavigationProp<RootStackParamList, 'Home'>;
-};
-
-const Home: React.FC<AppProps> = ({ navigation }): JSX.Element => {
+const Home = ({ navigation }) => {
     const theme = useTheme();
     const defaultStyles = styles(theme);
     const spinValue = new Animated.Value(0);
-    const securityHelper = useSecurityActions();
 
     Animated.loop(
         Animated.timing(spinValue, {
@@ -113,47 +75,16 @@ const Home: React.FC<AppProps> = ({ navigation }): JSX.Element => {
         outputRange: ['0deg', '360deg'],
     });
 
-    const changePassword = (): void => {
-        securityHelper.showChangePassword();
-    };
-
-    const logOut = (): void => {
-        LocalStorage.clearAuthCredentials();
-        securityHelper.onUserNotAuthenticated();
-    };
-
-    const menuItems: InfoListItemProps[] = [
-        { title: 'Change Password', IconClass: LockIcon, onPress: (): void => changePassword() },
-        { title: 'Log Out', IconClass: ExitToAppIcon, onPress: (): void => logOut() },
-    ];
-
     return (
         <>
             <Header
                 title={'Home Page'}
                 navigation={{
                     icon: MenuIcon,
-                    onPress: (): void => {
+                    onPress: () => {
                         navigation.openDrawer();
                     },
                 }}
-                actionItems={[
-                    {
-                        component: (
-                            <UserMenu
-                                menuItems={menuItems}
-                                avatar={
-                                    <Avatar.Text
-                                        label="UN"
-                                        size={40}
-                                        color={Colors.blue[500]}
-                                        style={{ backgroundColor: Colors.blue[50] }}
-                                    />
-                                }
-                            />
-                        ),
-                    },
-                ]}
             />
             <SafeAreaView style={defaultStyles.content}>
                 <ScrollView>
@@ -166,7 +97,7 @@ const Home: React.FC<AppProps> = ({ navigation }): JSX.Element => {
                         Welcome to PX <H4 color={'primary'}>Blue</H4>.
                     </H4>
                     <Body1 style={defaultStyles.subtitle}>
-                        Edit <Body1 style={defaultStyles.bold}>screens/home.tsx</Body1> and save to reload.
+                        Edit <Body1 style={defaultStyles.bold}>screens/home.js</Body1> and save to reload.
                     </Body1>
                     <Divider style={defaultStyles.divider} />
                     <OpenURLButton title={'PX Blue Documentation'} url={'https://pxblue.github.io/'} />
